extract Document from DocumentManager
